Fix application startup issues

+++ Backend/src/main/java/com/petconomy/controller/dto/MemberProfileDto.java
@@ -2,5 +2,5 @@ package com.petconomy.controller.dto;
 
 import java.math.BigDecimal;
 
-public record MemberProfileDto(int id, String username, String email, BigDecimal targetAmount) {
+public record MemberProfileDto(Long id, String username, String email, BigDecimal targetAmount) {
 }
